backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
Pillow==10.2.0
PyJWT==2.8.0
python-dotenv==1.0.0
gunicorn==21.2.0
gevent==24.2.1
bcrypt==4.1.2
